ryddet opp i 2D versjonene av koden

+++ 2D Org/main.h
@@ -14,5 +14,5 @@ struct Pixel{
 	Pixel(short k, short g):x(k), y(g){};
 };
 
-bool checkMaskNeighbours(int i, int j, int id, short res);
+bool checkMaskNeighbours(short i, short j, int id, short res);
